created sniffer library ... hopefully
